Deduplicate LfValueTranslation cache
changelog_begin
changelog_end
