+++ ledger/participant-integration-api/src/main/scala/platform/indexer/StandaloneIndexerServer.scala
@@ -9,7 +9,7 @@ import com.daml.ledger.resources.{Resource, ResourceContext, ResourceOwner}
 import com.daml.logging.{ContextualizedLogger, LoggingContext}
 import com.daml.metrics.Metrics
 import com.daml.platform.configuration.ServerRole
-import com.daml.platform.store.dao.events.LfValueTranslation
+import com.daml.platform.store.LfValueTranslationCache
 
 import scala.concurrent.ExecutionContext
 
@@ -18,7 +18,7 @@ final class StandaloneIndexerServer(
     config: IndexerConfig,
     servicesExecutionContext: ExecutionContext,
     metrics: Metrics,
-    lfValueTranslationCache: LfValueTranslation.Cache,
+    lfValueTranslationCache: LfValueTranslationCache.Cache,
 )(implicit materializer: Materializer, loggingContext: LoggingContext)
     extends ResourceOwner[Unit] {
 
@@ -70,7 +70,7 @@ final class StandaloneIndexerServer(
 
   private def startIndexer(
       indexer: RecoveringIndexer,
-      initializedIndexerFactory: ResourceOwner[JdbcIndexer],
+      initializedIndexerFactory: ResourceOwner[Indexer],
   )(implicit context: ResourceContext): Resource[Unit] =
     indexer
       .start(() => initializedIndexerFactory.flatMap(_.subscription(readService)).acquire())
