+++ ledger/participant-integration-api/src/main/scala/platform/store/IndexMetadata.scala
@@ -13,7 +13,6 @@ import com.daml.metrics.Metrics
 import com.daml.platform.ApiOffset
 import com.daml.platform.configuration.ServerRole
 import com.daml.platform.store.dao.JdbcLedgerDao
-import com.daml.platform.store.dao.events.LfValueTranslation
 import scalaz.Tag
 
 import scala.concurrent.{ExecutionContext, Future}
@@ -44,7 +43,7 @@ object IndexMetadata {
       eventsPageSize = 1000,
       servicesExecutionContext = executionContext,
       metrics = new Metrics(new MetricRegistry),
-      lfValueTranslationCache = LfValueTranslation.Cache.none,
+      lfValueTranslationCache = LfValueTranslationCache.Cache.none,
       enricher = None,
     )
