+++ ledger/participant-integration-api/src/main/scala/platform/store/dao/events/ContractsReader.scala
@@ -13,7 +13,7 @@ import com.daml.ledger.participant.state.index.v2.ContractStore
 import com.daml.logging.LoggingContext
 import com.daml.metrics.{Metrics, Timed}
 import com.daml.platform.store.Conversions._
-import com.daml.platform.store.DbType
+import com.daml.platform.store.{DbType, LfValueTranslationCache}
 import com.daml.platform.store.dao.DbDispatcher
 import com.daml.platform.store.dao.events.SqlFunctions.{H2SqlFunctions, PostgresSqlFunctions}
 import com.daml.platform.store.serialization.{Compression, ValueSerializer}
@@ -26,7 +26,7 @@ private[dao] sealed class ContractsReader(
     val committedContracts: PostCommitValidationData,
     dispatcher: DbDispatcher,
     metrics: Metrics,
-    lfValueTranslationCache: LfValueTranslation.Cache,
+    lfValueTranslationCache: LfValueTranslationCache.Cache,
     sqlFunctions: SqlFunctions,
 )(implicit ec: ExecutionContext)
     extends ContractStore {
@@ -90,7 +90,7 @@ private[dao] sealed class ContractsReader(
   )(implicit loggingContext: LoggingContext): Future[Option[Contract]] =
     // Depending on whether the contract argument is cached or not, submit a different query to the database
     lfValueTranslationCache.contracts
-      .getIfPresent(LfValueTranslation.ContractCache.Key(contractId)) match {
+      .getIfPresent(LfValueTranslationCache.ContractCache.Key(contractId)) match {
       case Some(createArgument) =>
         lookupActiveContractWithCachedArgument(readers, contractId, createArgument.argument)
       case None =>
@@ -129,7 +129,7 @@ private[dao] object ContractsReader {
       dispatcher: DbDispatcher,
       dbType: DbType,
       metrics: Metrics,
-      lfValueTranslationCache: LfValueTranslation.Cache,
+      lfValueTranslationCache: LfValueTranslationCache.Cache,
   )(implicit ec: ExecutionContext): ContractsReader = {
     def sqlFunctions = dbType match {
       case DbType.Postgres => PostgresSqlFunctions
